services/shared/template/requirements.txt
# requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# OpenTelemetry
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0
opentelemetry-exporter-otlp>=1.21.0

# Prometheus
prometheus-client>=0.19.0

# Async
httpx>=0.25.2
aiohttp>=3.9.0

# Kafka
aiokafka>=0.9.0

# Redis
redis>=5.0.0

# AI/ML
transformers>=4.36.0
torch>=2.1.0
openai>=1.6.0

# Utilities
python-dotenv>=1.0.0
structlog>=2.4.0
